Feature/#99: Helper 추가
