+++ Harubee-iOS/App/Sources/Presentation/Common/Views/TransactionInputView.swift
@@ -43,7 +43,7 @@ struct TransactionInputView: View {
         print("저장하기 Tap")
       }
       
-      NumberKeypadView(text: $expression) {
+      NumberKeypadView(expression: $expression) {
         print("Done")
       }
     }
