+++ Harubee-iOS/Sources/Helpers/Environments/EnvironmentKeys.swift
@@ -1,0 +1,13 @@
+//
+//  EnvironmentKeys.swift
+//  Harubee-iOS
+//
+//  Created by 이정동 on 11/3/24.
+//  Copyright © 2024 namudiEtc. All rights reserved.
+//
+
+import SwiftUI
+
+struct ButtonActionKey: EnvironmentKey {
+  static let defaultValue: () -> Void = {}
+}
